Add the Tails Debian repository that currently hosts a Jessie-compatible wiperam package.
This is meant to be temporary, as a way to quickly get something that one can
build and test. Before this is merged, the way to share that package between
Freepto and Tails shall be discussed (or rather retrieved from the previous
discussion on this topic, that may have happened on our development
mailing-lists). And then, possibly this commit will need to be reverted in
favour of another solution.